feat: add useShape and useShapeUUIDs hooks
