+++ frontend/iou/src/components/three/shape/ShapeRenderer.tsx
@@ -1,42 +1,18 @@
 import type {Vec3} from "@/hooks/workspace/workspaceTypes.ts";
 import InstancedVertexSpheres from "@/components/three/shape/InstancedVertexSpheres.tsx";
 import {useMemo, useState} from "react";
-import {Vector3} from "three";
 // TODO: Figure out why this import statement isn't happy
 import {ConvexGeometry} from "three/examples/jsm/geometries/ConvexGeometry";
 import type {ThreeEvent} from "@react-three/fiber";
 import {Edges} from "@react-three/drei";
+import {findClosestVertexId, vec3ToVector3} from "@/components/three/shape/vertex_helpers.ts";
+import ShapeWidget from "@/components/three/shape/ShapeWidget.tsx";
 
 export interface ShapeRendererProps {
   vertices: Vec3[],
   vertexColor?: string,
   baseColor?: string,
 }
-
-export function vec3ToVector3(vec3: Vec3): Vector3 {
-  return new Vector3(vec3[0], vec3[1], vec3[2]);
-}
-
-/**
- * Given a point and a set of vertices, gets the index of the closest vertex in vertices to the given point
- */
-export function findClosestVertexId(point: Vector3, vertices: Vec3[]): number {
-  let closest: number = 0;
-  let minDist = Infinity;
-
-  for (let i = 0; i < vertices.length; i++) {
-    const vertex = vec3ToVector3(vertices[i]);
-
-    const dist = vertex.distanceToSquared(point);
-    if (dist < minDist) {
-      minDist = dist;
-      closest = i;
-    }
-  }
-
-  return closest;
-}
-
 
 export default function ShapeRenderer(props: ShapeRendererProps) {
   const vertexColor = props.vertexColor ?? "blue";
@@ -57,6 +33,7 @@ export default function ShapeRenderer(props: ShapeRendererProps) {
   const onPointerOut = () => {
     setClosestVertexIds(null);
   }
+
 
   return (
     <group
