+++ frontend/iou/src/components/three/shape/vertex_helpers.ts
@@ -1,0 +1,26 @@
+import type {Vec3} from "@/hooks/workspace/workspaceTypes.ts";
+import {Vector3} from "three";
+
+export function vec3ToVector3(vec3: Vec3): Vector3 {
+  return new Vector3(vec3[0], vec3[1], vec3[2]);
+}
+
+/**
+ * Given a point and a set of vertices, gets the index of the closest vertex in vertices to the given point
+ */
+export function findClosestVertexId(point: Vector3, vertices: Vec3[]): number {
+  let closest: number = 0;
+  let minDist = Infinity;
+
+  for (let i = 0; i < vertices.length; i++) {
+    const vertex = vec3ToVector3(vertices[i]);
+
+    const dist = vertex.distanceToSquared(point);
+    if (dist < minDist) {
+      minDist = dist;
+      closest = i;
+    }
+  }
+
+  return closest;
+}
