+++ frontend/iou/src/hooks/workspace/useDimensions.ts
@@ -1,20 +1,6 @@
 import {useShallow} from "zustand/react/shallow";
 import type {Dimensions} from "@/hooks/workspace/workspaceTypes.ts";
-import {create} from "zustand/react";
-
-type DimensionsMutator = (dimensions: Dimensions) => void;
-
-export interface DimensionsStore {
-  setDimensions: DimensionsMutator;
-  dimensions: Dimensions;
-}
-
-export const useDimensionsStore = create<DimensionsStore>((set) => ({
-  dimensions: "3d",
-  setDimensions: (dimensions: Dimensions) => {
-    set({ dimensions: dimensions });
-  },
-}))
+import {useDimensionsStore, type DimensionsMutator} from "@/hooks/workspace/stores/useDimensionsStore.ts";
 
 export default function useDimensions(): [Dimensions, DimensionsMutator] {
   return useDimensionsStore(
